validation added

+++ OnlineTripManagementSystem/src/main/java/com/masai/service/TravelsServiceImpl.java
@@ -6,47 +6,81 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.masai.exceptions.LoginException;
 import com.masai.exceptions.TravelsException;
+import com.masai.exceptions.UserException;
+import com.masai.models.CurrentUserSession;
 import com.masai.models.Travels;
+import com.masai.repository.SessionDAO;
 import com.masai.repository.TravelsDAO;
 
 @Service
 public class TravelsServiceImpl implements TravelsService{
 	@Autowired
 	TravelsDAO travelsDao;//Jpa repositry
+
+	@Autowired
+	private SessionDAO sdao;
 	
 //=============================================Add Travels====================================================================
 	
 	@Override
-	public Travels addTravels(Travels travels) throws TravelsException {
+	public Travels addTravels(Travels travels, String uuid) throws TravelsException {
+		CurrentUserSession existingUser = sdao.findByUuid(uuid);
+		if(existingUser == null) {
+			throw new UserException("User not Logged In");
+		}
+		else if(existingUser.getUserType().equalsIgnoreCase("customer")) {
+			throw new LoginException("Access denied");
+		}
+		else {
 		Optional<Travels> tOpt=travelsDao.findById(travels.getTravelsId());//find travels if exist
 		if (tOpt.isPresent()) {
 			throw new TravelsException("This Travelservice is already available in database");
 		}
 		return travelsDao.save(travels);
+		}
 	}
 	
 //=============================================Update Travels details ==========================================================
 	
 	@Override
-	public Travels updateTravels(Travels travels) throws TravelsException {
+	public Travels updateTravels(Travels travels, String uuid) throws TravelsException {
+		CurrentUserSession existingUser = sdao.findByUuid(uuid);
+		if(existingUser == null) {
+			throw new UserException("User not Logged In");
+		}
+		else if(existingUser.getUserType().equalsIgnoreCase("customer")) {
+			throw new LoginException("Access denied");
+		}
+		else {
 		Optional<Travels> tOpt=travelsDao.findById(travels.getTravelsId());//find travels to update
 		if (!tOpt.isPresent()) {
 			throw new TravelsException("This Travelservice is not available in database to update");
 		}
 		return travelsDao.save(travels);
+		}
 	}
 	
 //=============================================Remove Travels====================================================================
 	
 	@Override
-	public Travels removeTravels(int travelId) throws TravelsException {
+	public Travels removeTravels(int travelId, String uuid) throws TravelsException {
+		CurrentUserSession existingUser = sdao.findByUuid(uuid);
+		if(existingUser == null) {
+			throw new UserException("User not Logged In");
+		}
+		else if(existingUser.getUserType().equalsIgnoreCase("customer")) {
+			throw new LoginException("Access denied");
+		}
+		else {
 		Optional<Travels> tOpt=travelsDao.findById(travelId);//find travels if exist with this Id
 		if (!tOpt.isPresent()) {
 			throw new TravelsException("This Travelservice is not present in database to remove");
 		}
 		travelsDao.delete(tOpt.get());
 		return tOpt.get();
+		}
 	}
 	
 //=============================================Search Travels by Id==============================================================
